Add CSV fixtures based on the test novel (#7)
* update test novel

* test data update

* update test path name

* add more test fixtures

* Unit test updates

* only emit full configs in debug mode

* update fixture paths

* update argument order

* update fixture paths

* remove settings.yaml from text example

* update skip print

* only inspect the latest output

* add json_parsing_llm function for handling cached outputs that need json parsing

* streamline top-level run usage

* json-parsing llm build updates

* streamline run_pipeline_with_config call pattern in examples

* skip data assertions on loaded_csv until the null-filtering for vectors lands

* fixture update

+++ python/graphrag/graphrag/index/storage/blob_pipeline_storage.py
@@ -47,16 +47,31 @@ class BlobPipelineStorage(PipelineStorage):
             self._container_name,
             self._path_prefix,
         )
+        self.create_container()
 
-        container_names = [
-            container.name for container in self._blob_service_client.list_containers()
-        ]
-        if container_name not in container_names:
-            self._blob_service_client.create_container(container_name)
+    def create_container(self) -> None:
+        """Create the container if it does not exist."""
+        if not self.container_exists():
+            container_name = self._container_name
+            container_names = [
+                container.name
+                for container in self._blob_service_client.list_containers()
+            ]
+            if container_name not in container_names:
+                self._blob_service_client.create_container(container_name)
 
     def delete_container(self) -> None:
         """Delete the container."""
-        self._blob_service_client.delete_container(self._container_name)
+        if self.container_exists():
+            self._blob_service_client.delete_container(self._container_name)
+
+    def container_exists(self) -> bool:
+        """Check if the container exists."""
+        container_name = self._container_name
+        container_names = [
+            container.name for container in self._blob_service_client.list_containers()
+        ]
+        return container_name in container_names
 
     def find(
         self,
